refactor: move tests/ to project root
Relocate graphene_mongo/tests/ → tests/ and convert all relative
`from ..` imports to absolute `from graphene_mongo.` imports.
Update Makefile and pyproject.toml paths accordingly.
