+++ tests/nodes_async.py
@@ -4,10 +4,10 @@ from graphene.relay import Node
 from . import models
 from . import types  # noqa: F401
 from .models import ProfessorMetadata
-from ..types import MongoengineObjectType
+from graphene_mongo.types_async import AsyncMongoengineObjectType
 
 
-class PublisherNode(MongoengineObjectType):
+class PublisherAsyncNode(AsyncMongoengineObjectType):
     legal_name = graphene.String()
     bad_field = graphene.String()
 
@@ -17,92 +17,92 @@ class PublisherNode(MongoengineObjectType):
         interfaces = (Node,)
 
 
-class ArticleNode(MongoengineObjectType):
+class ArticleAsyncNode(AsyncMongoengineObjectType):
     class Meta:
         model = models.Article
         interfaces = (Node,)
 
 
-class EditorNode(MongoengineObjectType):
+class EditorAsyncNode(AsyncMongoengineObjectType):
     class Meta:
         model = models.Editor
         interfaces = (Node,)
 
 
-class EmbeddedArticleNode(MongoengineObjectType):
+class EmbeddedArticleAsyncNode(AsyncMongoengineObjectType):
     class Meta:
         model = models.EmbeddedArticle
         interfaces = (Node,)
 
 
-class PlayerNode(MongoengineObjectType):
+class PlayerAsyncNode(AsyncMongoengineObjectType):
     class Meta:
         model = models.Player
         interfaces = (Node,)
         filter_fields = {"first_name": ["istartswith", "in"]}
 
 
-class ReporterNode(MongoengineObjectType):
+class ReporterAsyncNode(AsyncMongoengineObjectType):
     class Meta:
         model = models.Reporter
         interfaces = (Node,)
 
 
-class ParentNode(MongoengineObjectType):
+class ParentAsyncNode(AsyncMongoengineObjectType):
     class Meta:
         model = models.Parent
         interfaces = (Node,)
 
 
-class ChildNode(MongoengineObjectType):
+class ChildAsyncNode(AsyncMongoengineObjectType):
     class Meta:
         model = models.Child
         interfaces = (Node,)
 
 
-class ChildRegisteredBeforeNode(MongoengineObjectType):
+class ChildRegisteredBeforeAsyncNode(AsyncMongoengineObjectType):
     class Meta:
         model = models.ChildRegisteredBefore
         interfaces = (Node,)
 
 
-class ChildRegisteredAfterNode(MongoengineObjectType):
+class ChildRegisteredAfterAsyncNode(AsyncMongoengineObjectType):
     class Meta:
         model = models.ChildRegisteredAfter
         interfaces = (Node,)
 
 
-class ParentWithRelationshipNode(MongoengineObjectType):
+class ParentWithRelationshipAsyncNode(AsyncMongoengineObjectType):
     class Meta:
         model = models.ParentWithRelationship
         interfaces = (Node,)
 
 
-class ProfessorMetadataNode(MongoengineObjectType):
+class ProfessorMetadataAsyncNode(AsyncMongoengineObjectType):
     class Meta:
         model = ProfessorMetadata
         interfaces = (graphene.Node,)
 
 
-class ProfessorVectorNode(MongoengineObjectType):
+class ProfessorVectorAsyncNode(AsyncMongoengineObjectType):
     class Meta:
         model = models.ProfessorVector
         interfaces = (Node,)
 
 
-class ErroneousModelNode(MongoengineObjectType):
+class ErroneousModelAsyncNode(AsyncMongoengineObjectType):
     class Meta:
         model = models.ErroneousModel
         interfaces = (Node,)
 
 
-class BarNode(MongoengineObjectType):
+class BarAsyncNode(AsyncMongoengineObjectType):
     class Meta:
         model = models.Bar
         interfaces = (Node,)
 
 
-class FooNode(MongoengineObjectType):
+class FooAsyncNode(AsyncMongoengineObjectType):
     class Meta:
         model = models.Foo
         interfaces = (Node,)
