+++ tests/test_inputs.py
